Convert existing implementation files to Effect-based architecture
Replace existing files with Effect-based implementations instead of creating new ones:

- Convert src/index.ts to use Effect with dependency injection and tagged errors
- Transform src/database/connection.ts to Effect service with proper error handling
- Refactor src/services/embedding.ts from singleton class to Effect service
- Update src/services/ollama.ts from traditional async/await to Effect patterns
- Remove duplicate Effect-specific files (embedding-effect.ts, ollama-effect.ts, database.ts)
- Update src/layers/main.ts to reference converted service locations

All services now use Effect's functional programming approach:
- Tagged error types for domain-specific error handling
- Dependency injection via Effect Context system
- Composable operations through Effect.gen generators
- Type-safe error propagation throughout the service layer

Tests pass and TypeScript compilation succeeds with the converted implementation.
