+++ src/layers/main.ts
@@ -1,7 +1,7 @@
 import { Layer } from "effect"
-import { DatabaseServiceLive } from "../services/database"
-import { EmbeddingServiceLive } from "../services/embedding-effect"
-import { OllamaServiceLive } from "../services/ollama-effect"
+import { DatabaseServiceLive } from "../database/connection"
+import { EmbeddingServiceLive } from "../services/embedding"
+import { OllamaServiceLive } from "../services/ollama"
 
 export const AppLayer = Layer.mergeAll(
   DatabaseServiceLive,
