+++ src/services/ollama.ts
@@ -1,62 +1,77 @@
+import { Context, Effect, Layer } from "effect"
 import { Ollama } from "ollama"
+import { OllamaConnectionError, OllamaModelError } from "../errors/ollama"
 
-const ollama = new Ollama({
-  host: "http://localhost:11434",
+export interface OllamaService {
+  readonly generateEmbedding: (
+    text: string,
+    modelName?: string
+  ) => Effect.Effect<number[], OllamaModelError>
+  readonly isModelAvailable: (
+    modelName?: string
+  ) => Effect.Effect<boolean, OllamaConnectionError>
+  readonly pullModel: (
+    modelName?: string
+  ) => Effect.Effect<void, OllamaModelError>
+}
+
+export const OllamaService = Context.GenericTag<OllamaService>("OllamaService")
+
+const make = Effect.gen(function* () {
+  const ollama = new Ollama({
+    host: "http://localhost:11434",
+  })
+
+  const generateEmbedding = (text: string, modelName = "embeddinggemma:300m") =>
+    Effect.tryPromise({
+      try: async () => {
+        const response = await ollama.embeddings({
+          model: modelName,
+          prompt: text,
+        })
+        return response.embedding
+      },
+      catch: (error) =>
+        new OllamaModelError({
+          message: `Failed to generate embedding using model ${modelName}`,
+          modelName,
+          cause: error,
+        }),
+    })
+
+  const isModelAvailable = (modelName = "embeddinggemma:300m") =>
+    Effect.tryPromise({
+      try: async () => {
+        const models = await ollama.list()
+        return models.models.some((model) => model.name.includes(modelName))
+      },
+      catch: (error) =>
+        new OllamaConnectionError({
+          message: "Failed to check model availability",
+          cause: error,
+        }),
+    })
+
+  const pullModel = (modelName = "embeddinggemma:300m") =>
+    Effect.tryPromise({
+      try: async () => {
+        console.log(`Pulling model ${modelName}...`)
+        await ollama.pull({ model: modelName })
+        console.log(`Model ${modelName} pulled successfully`)
+      },
+      catch: (error) =>
+        new OllamaModelError({
+          message: `Failed to pull model ${modelName}`,
+          modelName,
+          cause: error,
+        }),
+    })
+
+  return {
+    generateEmbedding,
+    isModelAvailable,
+    pullModel,
+  } as const
 })
 
-export interface EmbeddingResponse {
-  embedding: number[]
-}
-
-export class OllamaService {
-  private static instance: OllamaService
-
-  private constructor() {}
-
-  static getInstance(): OllamaService {
-    if (!OllamaService.instance) {
-      OllamaService.instance = new OllamaService()
-    }
-    return OllamaService.instance
-  }
-
-  async generateEmbedding(
-    text: string,
-    modelName = "embeddinggemma:300m"
-  ): Promise<number[]> {
-    try {
-      const response = await ollama.embeddings({
-        model: modelName,
-        prompt: text,
-      })
-
-      return response.embedding
-    } catch (error) {
-      console.error("Failed to generate embedding:", error)
-      throw new Error(
-        `Failed to generate embedding from Ollama using model ${modelName}`
-      )
-    }
-  }
-
-  async isModelAvailable(modelName = "embeddinggemma:300m"): Promise<boolean> {
-    try {
-      const models = await ollama.list()
-      return models.models.some((model) => model.name.includes(modelName))
-    } catch (error) {
-      console.error("Failed to check model availability:", error)
-      return false
-    }
-  }
-
-  async pullModel(modelName = "embeddinggemma:300m"): Promise<void> {
-    try {
-      console.log(`Pulling model ${modelName}...`)
-      await ollama.pull({ model: modelName })
-      console.log(`Model ${modelName} pulled successfully`)
-    } catch (error) {
-      console.error(`Failed to pull model ${modelName}:`, error)
-      throw new Error(`Failed to pull model ${modelName}`)
-    }
-  }
-}
+export const OllamaServiceLive = Layer.effect(OllamaService, make)
